Skyline: tweak this script to remove spurious warning when PrtocolBuffers has already been cleaned

+++ pwiz_tools/Skyline/CleanSkyline.bat
@@ -40,6 +40,6 @@ IF EXIST ..\Shared\MSGraph\bin rmdir /s /q ..\Shared\MSGraph\bin
 IF EXIST ..\Shared\MSGraph\obj rmdir /s /q ..\Shared\MSGraph\obj
 IF EXIST ..\Shared\Crawdad\obj rmdir /s /q ..\Shared\Crawdad\obj
 IF EXIST ProtocolBuffers\tmp rmdir /s /q ProtocolBuffers\tmp
-del /q ProtocolBuffers\GeneratedCode\*.cs
+IF EXIST ProtocolBuffers\GeneratedCode\*.cs del /q ProtocolBuffers\GeneratedCode\*.cs
 
 popd
